Refactor with icons, interface and description
