+++ src/components/ui/icons/check-icon.tsx
@@ -1,0 +1,11 @@
+export const CheckIcon = () => (
+  <svg viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M11.6666 3.5L5.24992 9.91667L2.33325 7"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);
