+++ src/components/ui/icons/index.ts
@@ -1,3 +1,5 @@
 import { UserIcon } from './user-icon';
+import { CheckIcon } from './check-icon';
+import { MinusIcon } from './minus-icon';
 
-export { UserIcon };
+export { UserIcon, CheckIcon, MinusIcon };
